fix: oauth csp requirements
chrome doesn't seem to like the form-action csp directive. It seems extrenious, and is not listed
in https://developer.mozilla.org/en-US/docs/Web/HTTP/Reference/Headers/Content-Security-Policy#scheme-source

AND it also did want a connect-src in order to report RUM and csp-reports to datadog so that was added.

I've also added another test target called 'smoke_test' where we can test in integration
easily using playwright. The test for this was added there because getting it running locally
is difficult, since webpack js bundles are not
loaded during test. That'll probably need to be
fixed in the future, but coupled with the requirement
to have a client and a secret, this just made more sense to
put over there.

Refs: https://new-hippo.atlassian.net/browse/HC-43