feat: ensure real webpack is used for e2e tests
Ensuring real webpack is used means that yarn run build and
collectstatic must run to ensure up to date bundle is available.

This is accomplished by running a django management command as part of
the pytest collection, and then also overriding the webpack settings for
the tests marked as e2e.