feat: assign task and encounter perms on claim patient
https://new-hippo.atlassian.net/browse/HC-61

When a user accepts an invite and claims a patient, assign the user
default permissions for the encounter and tasks created for that
patient.

Leaving out migrations for the sake of time since this can be remedied
by simply inviting a new patient. If people really want a backfilling
migration I'm happy to include one.