feat: try harder to stay in encounter-land
providers should only be looking at a patient outside the context of an encounter in the moment before deciding to create a new encounter.

if there's already an encounter, prefer to look at it

there's a bunch of clean up here removing duplicate routes; we only need one patient-add view, etc.

https://new-hippo.atlassian.net/browse/HC-160