chore: reset base-unit to default in surveyJS theme
https://new-hippo.atlassian.net/browse/HC-202

`base-unit` was being used for more than just margins and spacing. It
was affecting the width of elements and event the text font size. The
easiest path forward for now is to reset it back to default as opposed
to introducing a bunch of exceptions.

We can revisit in the future.