fix: health summary survives page refresh
Health summary survives page refresh by persisting the summary
to the new PersonalSummary model instead of the in-mem cache. Chat context
is updated with most recent health summary generated within the last 24hrs.

Using the cache here does not work since its an in-memory cache. This means
that the cache gets wiped on every deploy, resulting in the bug reported.

The bug doesn't show itself during local development until the server is
restarted.

Future caution for using the in-memory cache, from the docs:
>Note that each process will have its own private cache instance, which means
no cross-process caching is possible.

This means cache access between the app and the workers is not possible with our current
configuration, so don't do that.
https://docs.djangoproject.com/en/5.2/topics/cache/#local-memory-caching

https://new-hippo.atlassian.net/browse/HC-292